requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
mysql-connector-python==8.2.0
Werkzeug==3.0.1
Flask-SocketIO==5.3.6
google-genai==0.3.0
pydantic>=2.12.5
python-dotenv==1.0.1
